services/recurring-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dateutil==2.8.2
rrule==0.2.12
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
dapr-dev==0.0.0.dev23
